+++ ts/packages/json-schema-to-zod/eslint.config.mjs
@@ -1,6 +1,6 @@
 module.exports = {
-    extends: ['../../.eslintrc.cjs'],
-    // parserOptions: {
-    //     project: ['./tsconfig.json'],
-    // },
-}
+  extends: ['../../.eslintrc.cjs'],
+  // parserOptions: {
+  //     project: ['./tsconfig.json'],
+  // },
+};