+++ ts/packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @composio/core
 
+## 0.2.3
+
+### Patch Changes
+
+- f88ab99: Fix: SDK telemetry methods and request headers
+
+## 0.2.2
+
+### Patch Changes
+
+- cfc2c50: Update zod version to 4
+- Updated dependencies [cfc2c50]
+  - @composio/json-schema-to-zod@0.1.18
+
 ## 0.2.1
 
 ### Patch Changes