+++ ts/packages/providers/vercel/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @composio/vercel
 
+## 0.2.15
+
+### Patch Changes
+
+- Updated dependencies [f88ab99]
+  - @composio/core@0.2.3
+
+## 0.2.14
+
+### Patch Changes
+
+- cfc2c50: Update zod version to 4
+- Updated dependencies [cfc2c50]
+  - @composio/core@0.2.2
+
 ## 0.2.13
 
 ### Patch Changes